// edgaras.com-2026-03-20/lib/types/supabase.ts
export interface Database {
  public: {
    Tables: {
      articles: {
        Row: {
          created_at: string
          description: string
          markdown: string
          published_at: string
          slug: string
          status: string
          tags: string | null
          title: string
          updated_at: string
        }
        Insert: {
          created_at?: string
          description: string
          markdown: string
          published_at?: string
          slug: string
          status: string
          tags?: string | null
          title: string
          updated_at?: string
        }
        Update: {
          created_at?: string
          description?: string
          markdown?: string
          published_at?: string
          slug?: string
          status?: string
          tags?: string | null
          title?: string
          updated_at?: string
        }
        Relationships: []
      }
      bookmarks: {
        Row: {
          created_at: string | null
          id: number
          link: string
          tags: string[] | null
          title: string | null
        }
        Insert: {
          created_at?: string | null
          id?: number
          link: string
          tags?: string[] | null
          title?: string | null
        }
        Update: {
          created_at?: string | null
          id?: number
          link?: string
          tags?: string[] | null
          title?: string | null
        }
        Relationships: []
      }
      newsletters: {
        Row: {
          created_at: string
          html: string | null
          id: number
          status: string
          subject: string | null
          text: string | null
        }
        Insert: {
          created_at?: string
          html?: string | null
          id?: number
          status?: string
          subject?: string | null
          text?: string | null
        }
        Update: {
          created_at?: string
          html?: string | null
          id?: number
          status?: string
          subject?: string | null
          text?: string | null
        }
        Relationships: []
      }
      posts: {
        Row: {
          created_at: string
          id: string
          markdown: string
          status: string
        }
        Insert: {
          created_at?: string
          id?: string
          markdown: string
          status: string
        }
        Update: {
          created_at?: string
          id?: string
          markdown?: string
          status?: string
        }
        Relationships: []
      }
    }
    Views: {
      bookmark_tags_all: {
        Row: {
          count: number | null
          name: string | null
        }
        Relationships: []
      }
    }
    Functions: {
      generate_random_19_digit_id: {
        Args: Record<PropertyKey, never>
        Returns: number
      }
      test_send_newsletter_batch: {
        Args: Record<PropertyKey, never>
        Returns: undefined
      }
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}

export type Tables<
  PublicTableNameOrOptions extends
    | keyof (Database["public"]["Tables"] & Database["public"]["Views"])
    | { schema: keyof Database },
  TableName extends PublicTableNameOrOptions extends { schema: keyof Database }
    ? keyof (Database[PublicTableNameOrOptions["schema"]]["Tables"] &
        Database[PublicTableNameOrOptions["schema"]]["Views"])
    : never = never
> = PublicTableNameOrOptions extends { schema: keyof Database }
  ? (Database[PublicTableNameOrOptions["schema"]]["Tables"] &
      Database[PublicTableNameOrOptions["schema"]]["Views"])[TableName] extends {
      Row: infer R
    }
    ? R
    : never
  : PublicTableNameOrOptions extends keyof (Database["public"]["Tables"] &
      Database["public"]["Views"])
  ? (Database["public"]["Tables"] &
      Database["public"]["Views"])[PublicTableNameOrOptions] extends {
      Row: infer R
    }
    ? R
    : never
  : never
// ...
export type Enums<
  PublicEnumNameOrOptions extends
    | keyof Database["public"]["Enums"]
    | { schema: keyof Database },
  EnumName extends PublicEnumNameOrOptions extends { schema: keyof Database }
    ? keyof Database[PublicEnumNameOrOptions["schema"]]["Enums"]
    : never = never
> = PublicEnumNameOrOptions extends { schema: keyof Database }
  ? Database[PublicEnumNameOrOptions["schema"]]["Enums"][EnumName]
  : PublicEnumNameOrOptions extends keyof Database["public"]["Enums"]
  ? Database["public"]["Enums"][PublicEnumNameOrOptions]
  : never